.history/requirements_20231018110556.txt
feedparser
tenacity
flask
requests
PyYAML